intel: ocl: rnn: create cell fusion function for lbr gru, lstm, rnn
