+++ src/gpu/intel/ocl/rnn/cell_gru_lbr.cpp
@@ -18,7 +18,7 @@
  * Cell execution GRU with linear before reset
  */
 
-#include "gpu/intel/ocl/rnn/rnn_grid.hpp"
+#include "gpu/intel/ocl/rnn/simple_cell_fusion.hpp"
 
 namespace dnnl {
 namespace impl {
@@ -38,18 +38,32 @@ cell_execution_sig((simple_rnn_common_t<aprop>::cell_execution_gru_lbr)) {
     const ocl_conf_t &ocl_conf = this->pd()->ocl_conf;
     const rnn_offsets_t &offsets = this->pd()->off;
 
+    const bool use_cell = ocl_conf.cell_comp.is_enabled;
+
+    strides_t<4> user_layer_strides {[&]() {
+        auto s = user_data.src_layer_strides(dir);
+        return strides_t<4> {0, 0, s[0], s[1]};
+    }()};
+
     auto cell_layer = !rnn.copy_src_layer && lay == 0
             ? user_data.src_layer(dir, iter)
             : workspace.states(lay - 1, dir, iter);
+    auto &cell_layer_strides = !rnn.copy_src_layer && lay == 0
+            ? user_layer_strides
+            : workspace.states_strides();
+    auto cell_iter = workspace.states(lay, dir, iter - 1);
+    auto cell_iter_strides = workspace.states_strides();
+
     auto gemm_cell_layer_fwd = !rnn.copy_src_layer && lay == 0
             ? gemm_layer_fwd_src
             : gemm_layer_fwd;
     auto gemm_diff_wei_cell_layer = !rnn.copy_src_layer && lay == 0
             ? gemm_diff_wei_layer_src
             : gemm_diff_wei_layer;
-    auto cell_iter = workspace.states(lay, dir, iter - 1);
 
     auto scratch_gates = scratch.gates(iter);
+    strides_t<2> scratch_gates_strides
+            = {scratch.calc_off_gates(1), rnn.scratch_gates_ld};
     auto &scratch_cell = scratch.cell() ? *scratch.cell()
                                         : memory_storage_t::empty_storage();
 
@@ -58,17 +72,26 @@ cell_execution_sig((simple_rnn_common_t<aprop>::cell_execution_gru_lbr)) {
 
     if (aprop == prop_kind::forward) {
         // call made when cell execution is enabled
-        if (!rnn.merge_gemm_layer)
+        if (!rnn.merge_gemm_layer && !rnn.cell_fusion.gemm_layer)
             CHECK(gemm_primitive(engine, ctx, wei_layer, cell_layer,
                     scratch_gates, gemm_cell_layer_fwd));
 
-        CHECK(gemm_primitive(
-                engine, ctx, wei_iter, cell_iter, scratch_cell, gemm_iter_fwd));
+        if (!rnn.cell_fusion.gemm_iter)
+            CHECK(gemm_primitive(engine, ctx, wei_iter, cell_iter, scratch_cell,
+                    gemm_iter_fwd));
 
-        CHECK((this->*elemwise_gru_lbr)(ctx, dir, lay, iter, rnn.dhc, rnn.mb, 1,
-                user_data, workspace, scratch_gates, {}, scratch_cell, {}, {},
-                {}, 0, tm_scales, diff_bias));
+        if (!use_cell) {
+            CHECK((this->*elemwise_gru_lbr)(ctx, dir, lay, iter, rnn.dhc,
+                    rnn.mb, 1, user_data, workspace, scratch_gates, {},
+                    scratch_cell, {}, {}, {}, 0, tm_scales, diff_bias));
+        } else {
 
+            CHECK(compute_cell_fwd(ctx, kernels_[kernel_id::cell_fwd], lay, dir,
+                    iter, workspace, user_data, wei_layer, wei_iter, cell_layer,
+                    cell_layer_strides, cell_iter, cell_iter_strides,
+                    scratch_gates, scratch_gates_strides, scratch_cell,
+                    pd()->desc()->alpha, tm_scales, rnn, ocl_conf, offsets));
+        }
     } else {
         auto diff_states_iter = scratch.diff_states(lay, dir, 0, iter + 1);
         auto diff_states_layer
